+++ tests/testthat/test_setwrapper.R
@@ -6,10 +6,10 @@ test_that("construction",{
   expect_error(SetWrapper$new(), "SetWrapper is an abstract")
 })
 
-test_that("accessors",{
-  u = Set$new(1) + Interval$new(3, 4)
-  expect_equal(u$class, "numeric")
-})
+# test_that("accessors",{
+#   u = Set$new(1) + Interval$new(3, 4)
+#   # expect_equal(u$class, "numeric")
+# })
 
 test_that("equals",{
   u = Set$new(1) + Interval$new(3, 4)