+++ NEWS.md
@@ -6,6 +6,11 @@
 - Improved printing of `SpecialSet`s when `zero == TRUE`
 - Added `UniversalSet` for the set containing all elements
 - Changed default `universe` of sets to `UniversalSet`
+- Coercions now error instead of producing a message when they fail
+- On construction, `Set`s no longer guess the set class, instead an extra `class` argument is added to give a set the `typed` property
+- The internal `Set` structure is slightly changed so that set elements are now stored in lists by default, which is only changed if the set is `typed`
+- Added `element` argument to `Set` constructor, which takes a `list`. This is more efficient if passing lists of lists or lists of multiple types, and in line with the `FuzzySet` constructor
+- Improved printing of `ConditionalSet`s
 
 # set6 0.1.0